requirements.txt
python-dotenv==0.21.0
requests~=2.28.1
loguru~=0.6.0
pyTelegramBotAPI~=4.8.0
